improved objective
Before, the objective did not account for the fact that covariances can be negative. Now we take the absolute value of the covariances to force them towards 0 rather than towards a negative number.

Now, the variances in the objective are weighted to be just important as the covariances. Without this weighting, the network seems to produce very correlated outputs with pretty much 0 variance, bringing the covariances *and* variances close to 0 (thus bringing the objective close to 0, when it could be negative).